refactor: extract tocode
